usb-rndis: add config for g_ether.ko solution
i.MX6 and Vybrid modules use the g_ether driver for RNDIS functionality.
This commit uses a systemd service to load the kernel module, configures the new
usb0 NIC and starts a dhcp server to serve an IP address to a USB host.

The Tegras use the android gadget driver. The configuration remains unchanged.