libusbg: enable USB Gadgets after Connman to avoid interference
It seems that Connman disables ALL interfaces on startup, even if
they are listed in NetworkInterfaceBlacklist. This leads to the
following situation:

networkd starts...
usb0 gets created...
networkd sets the device up (event triggered)
connman starts... and sets the device down!

networkd does not restart the device in this situation. To avoid
the situation, create the USB Gadget functions after connman. It
seems that if connman is up and running, it does not touch the
device anymore.