libusbg: make sure USB RNDIS Gadget is started before networkd
Despite the reordering of when to load the USB Gadget configuration
with commit fadb16f153aa ("libusbg: enable USB Gadgets after Connman
to avoid interference") the network device ended up to be down after
startup. It seems that there is a race condition in systemd-networkd
which triggers especially in combination with USB Gadget, a issue on
the systemd issue tracker has been created:
https://github.com/systemd/systemd/issues/1784

There have been various changes related to link handling between
systemd 219 and the latest version 227. It is possible that the issue
is already solved.

Until this has been resolved, explicitly load the USB Gadget before
starting systemd-networkd seems to get around the issue.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>