hack around Qt hacks
Ok, hold on, this is a hack to work around a Qt/Qwt specific hack...
When using the LinuxInput keyboard driver (QWS_KEYBOARD="LinuxInput..)
the Qwt LinuxInput driver at src/gui/embedded/qkbdlinuxinput_qws.cpp
tries to detect whether stdin is a terminal or not (using isatty). It
seems that this is true in our minimalistic init case, which leads the
Qt Qws LinuxInput driver to trigger a workaround which reconfigures the
currently running TTY to defaults (baud rate 9600!!).

The idea behind is that if it is a real tty, and we use the LinuxInput
driver, a Ctrl+C might kill the Qt application (instead copy something).

How to work around? Probably correct would be to change Qt Qws LinuxInput
driver to detect serial terminals... A lot simpler is to just set stdin
to /dev/zero, with what Qt Qwt LinuxTermial does not trigger the work
around (or at least not on our serial terminal device).

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>