image_type_tezi: add custom image type for Toradex Easy Installer
The Toradex Easy Installer is a Linux based installer for images. This
commit adds a new image type which allows to build images flashable with
this installer directly with OpenEmbedded.

The installer needs a JSON based image description (image.json) which will
be generated using a custom Python step. The IMAGE_CMD step then creates
a bootfs tar.xz and creates a tar with the bootfs, rootfs, U-Boot and SPL,
image.json and some more metadata files. This tarball can be extracted
on a SD card, USB flah drive or HTTP server and consumed by the Toradex
Easy Installer.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>