apalis/colibri-imx6/t30/tk1: read ahead cache configuration
Configure the read_ahead cache size to optimal value, for the
mmcblk devices (eMMC and SD card).

The cache values are module specific, tests were run on the pre-built
image (BSP V2.7 Beta 1). The measurement of read speed for a range of
cache sizes (16kB to 256MB) was made using hdparm -t.

Signed-off-by: Leonardo Graboski Veiga <leogveiga@gmail.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>