do not add image_type_tezi_run unconditionally
So far we added image_type_tezi_run image class unconditionally in
the tezi distro. However, we do not use the FIT image format for the
tester, hence we do not want to build that image format.

Also make sure that dependencies of the image format (e.g. the boot
scripts generated by u-boot-distro-boot) are properly declared inside
the image class.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>