define UBOOT_BINARY if not defined already
We use UBOOT_BINARY to pack the final zip file since this is
the only reliable variable for NAND based devices. Define the
variable if it is not already defined the same way as
uboot-config.bbclass defines it. While at it, also use the
defined variable when creating the json file for eMMC based
devices.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>