image_type_tezi_run: do not add prepare script
The runtime check is flawed since a previous regular image build
(which has an empty sample prepare.sh script) creates a prepare.sh
script in the deploy folder. However, the Tezi image paket does not
include that script, hence it will be missing at runtime but
referenced from the image.json file.

So far we have no need for a prepare script in Tezi, hence just
don't reference a script in the Tezi image class.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>