image_type_tezi_run: do not add marketing.tar
Like the prepare.sh script, the runtime check for the marketing.tar
file in the deployment folder is flawed since a previous regular
image build creates a marketing.tar too. However, the Tezi image
paket does not include a marketing.tar, hence it will be missing at
runtime but referenced from the image.json file.

Note that this only showed when downloading the image from a HTTP
server, from a local media Tezi seems to silently ignore when the
tar file is missing.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>