libusbgx: fix usb product id calculation
The Toradex product ID from the config block is decimal coded
ASCII with leading zeros. In C and bash, leading zeros means octal
number... This leads to wrong calculation of the USB product ID
and to errors if the product ID contains the number 8 or 9 (since
those are not valid under the octal numeric system).

Fix both issues by explicitly force decimal system for the Toradex
product ID.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>