qt-tezi: use TTY keyboard driver again
Unfortunately, the LinuxInput keyboard driver in Qt seems not
to detect keyboards well: On Colibri iMX6 the mouse/keyboard
combo did not work anymore, and on an Apalis using a dedicated
USB mouse and keyboard seemed not to work. Also hotplug seems
not to work.

The TTY driver has caused Qt UI hangs. However, in quick tests
with the new Kernel 4.1 they have not been observed, so lets
give the TTY driver another try.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>