imx-kobs: extend with support to specify MTD size
The imx-kobs utility expects a linear MTD device to write
the BCB/DBBT and the firmware. However, our default MTD
layout usually splits the BCB/DBBT area and the boot loader
into separate MTD partitions.

Furthermore, kobs-ng divides the size after the BCB/DBBT in
two and writes the first firmware (boot stream) at the
beginning of the flash and the second boot stream in the
middle of the flash...

By using the Linux configuration CONFIG_MTD_PARTITIONED_MASTER
we have a linear /dev/mtd0 of the whole NAND flash. And with
this patch we add support to specify the MTD size. By specifying
the size of the BCB and the two U-Boot partitions (0x380000)
kobs-ng will use exectly the expected layout:
- BCB/DBBT at the beginning up to 0x80000
- First boot stream at 0x00080000
- Second boot stream at 0x00200000

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>