image_type_tezi: write uEnv.txt as default U-Boot environment
Use the "u_boot_env" property to write uEnv.txt to flash
by default. With that the Toradex Easy Installer will make
sure that the default environment of the shipped U-Boot is
written on flash.

Especially since with 2.8b2 we transition to using zImage,
this change will make sure that a correct U-Boot environment
is written on flash to boot the zImage.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>