u-boot-distro-boot: Use VESA VGA on parallel RGB
It seems that even asking for 640x480M@60 does not lead to proper
VESA VGA timing. The pixel clock is at 23.175MHz which did not work
with some screens.

Use 640x480@60, which chooses the VGA resolution from modedb. This
seems to lead to proper VESA VGA display timings and work well on the
EDT displays too.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>