correct zip-native dependency for tezi-run target
tezi-run fails to zip the image at the end of the build process since
zip-native is not correctly pulled into the environment. It requires to be set
in the DEPENDS list. Additionally, oe rocko now requires inter-task
dependencies to specify the task as <recipe-name>:<task>. Correct those for the
tezi-run class.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>