u-boot-distro-boot: fix incorrect variable in boot.cmd
When using Distro Boot, U-Boot searches through all partitions in each device
through the $distro_bootpart dummy variable. The current $bootpart variable used
in our scripts is incorrect. Serendipitously, an unset variable in the
partition number of the load command makes it default to the first partition,
which makes Distro Boot always end up loading the first valid partition, but
never a boot-flagged partition other than the first one. This needs to be fixed
to comply with the Distro Boot specification.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>