weston: fix target rpath issue
There is a flaw in meta-openembedded/meta-oe/classes/meson.bbclass,
introduced by commit 6fb9a9e7:
[ meson: set native tool flag variables ]

since meson slurps any set CFLAGS, CXXFLAGS, LDFLAGS, and CPPFLAGS from
the environment and injects them into the build, this means that we are
seeing native CFLAGS, CXXFLAGS, LDFLAGS, and CPPFLAGS in the target
build.

And this was fixed in OE's meson.bbclass, by commit de7ae028:
[ meson: export native env only for native build ]

We cherry pick part of that fix to weston recipe, since it is the only
recipe that inheriting meson.bbclass.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>