libx11: split out libx11-locale
So far libx11-locale is also being required by some non-X11 recipes,
like weston -> libxkbcommon -> libx11-locale. But the libx11-locale
could not build without x11 distro feature enabled. This is not
reasonable.

To fix it, we split out libx11-locale from libx11 to be a standalone
recipe, and drop the X11 dependencies, this makes it to be able to
build without x11 distro feature.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>