image_type_tezi_run.bbclass: remove boot-sdp.scr from the deploy tar
imx8's do not need a boot-sdp.scr in the deployed tarball, as this
script is now embedded in the imx8 horrendous boot container.
Unfortunately this also means that customers cannot use that script to
manipulate the u-boot environment (they can still use teziargs), but
this shouldn't be a major issue in recovery.

Signed-off-by: Gerard Salvatella <gerard.salvatella@toradex.com>