tezi.conf: drop _mx8 overrides when setting HOSTTOOLS_NONFATAL
It has a obvious drawback to set HOSTTOOLS_NONFATAL for a specific
machine that it will impact the base hash of tasks for different
machines hence will lead to unnecessary rebuild.

Adding sha384sum to HOSTTOOLS_NONFATAL is not harmful to other machines
as well since it's 'nofatal', so we can drop _mx8 overrides safely.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>