pixman: add aarch64 assembler support
It seems that even the new version of Pixman breaks RDP screen
sharing support in Weston. It seems that the the fallback
implementation of some pixel operations are broken. Adding a
patch implementing aarch64 NEON acceleration seems to provide
working implementation for aarch64 and makes RDP screen
sharing work on Apalis iMX8.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>