image_type_tezi_run.bbclass: refactor tezirunimg based upon sstate
This patch mainly aims to handle the tezi.itb, json files, and Tezi
installer tarball in sstate.

To achieve that, do_image_tezirunimg needs to run before
do_image_complete, and do_assemble_fitimage should be dropped.

To be able to generate tezi.itb with the squashfs image in
${IMGDEPLOYDIR}, we make a temp ${DEPLOY_DIR_IMAGE}/tezi-XXXXXX.its and
substitute the relative path of squashfs image './' with the absolute
path ${IMGDEPLOYDIR} before calling mkimage.

We also set RM_WORK_EXCLUDE += "${PN}", this ensures the squashfs image
could be accessed in do_image_tezirunimg triggered by later bitbake
instances.

To be noticed: after this change, the build should be started from a
clean build directory since the original tezi.itb is not being
controlled by sstate which will lead to a following conflict:
| do_image_complete: The recipe tezi-run is trying to install files
| into a shared area when those files already exist:
|  /toradex/tezi-thud-build/build/deploy/images/apalis-imx8/tezi.itb
|    (not matched to any task)

If you do not want to build from a clean directory, then just simply
remove that conflict file and restart a build, which would also work.

Signed-off-by: Ming Liu <ming.liu@toradex.com>