base-passwd: Add kvm group
commit 76caed136b819522b908a2bee7b81103bd2ea7e4 OE upstream.

Although base-passwd in OE is somewhat outdated, upgrading
to a newer version is not going to solve eudev warnings about
missing groups during boot; input/shutdown/kvm are still not
listed in groups.master. The reason for this is that
Debian uses systemd, which will automatically create
missing groups(systemd-sysusers). In a sysvinit+eudev
configuration you instead get a warning printed to
the console:

udevd[<pid>]: specified group 'kvm' unknown

Signed-off-by: Ming Liu <ming.liu@toradex.com>