force HDMI connector to be always connected
Force the HDMI connector to be considered as always connected. This
makes sure that Tezi gets started even when no HDMI is plugged in.
Without this, if there is no HDMI monitor connected, Tezi just exits
with:
  / # cat /var/volatile/tezi.log
  Running on a compositor with no screens is not supported

It also prevents a race condition where sometimes the monitor is not
detected at startup and only some seconds later.

Related-to: #58812
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>