linux-toradex-mainline-5.4: fix defconfig for colibri-imx7
Between 5.2 and 5.4 we changed the meta-toradex-bsp-common defconfig
from more generic mx7 to more specific colibri-imx7 and
colibri-imx7-emmc (even though they are both the same). Anyway, for Tezi
to pick up the proper one again rename mx7 to colibri-imx7 as well (the
colibri-imx7-emmc one is not used in Tezi as Tezi is "generic".

While at it also generalise the FILESEXTRAPATHS stuff to use ${PN}.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>