weston: backend-vnc: treat shift as a hint
Ignore shift state as per RFC6143 Section 7.5.4 [1]:

The "shift state" (i.e., whether either of the Shift keysyms is
down) should only be used as a hint when interpreting a keysym.
For example, on a US keyboard the '#' character is shifted, but on
a UK keyboard it is not.  A server with a US keyboard receiving a
'#' character from a client with a UK keyboard will not have been
sent any shift presses.  In this case, it is likely that the
server will internally need to simulate a shift press on its local
system in order to get a '#' character and not a '3'.

[1] https://tools.ietf.org/html/rfc6143#section-7.5.4

Signed-off-by: Igor Opaniuk <igor.opaniuk@toradex.com>