image_type_tezi_run: handle emmc being mmcblk2
Note that we currently do not want to use the stable symlinks to the
eMMC for modules were older version of the Toradex Easy Installer exist.
The new Toradex Easy Installer could be installed by an older
one which can not cope with the /dev/emmc symlink.

For Verdin iMX8MP this is the first version and thus we can directly use
the stable symlink name pointing to the eMMC.

Related-to: TEI-750
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>