u-boot-distro-boot: adjust boot.cmd for native hdmi on verdin-imx8mp
Unfortunately, the previous HDMI display adapter resolution of
1024x768@60 does not seem suitable for native HDMI. The lowest common
denominator seems to be 1280x720 which boils down to the following
vidargs:

video=HDMI-A-1:1280x720-16@60D video=HDMI-A-2:1280x720-16@60D

Note that the D means it is a digital display that is assumed to always
be connected. This avoids any kind of negotiation issues. But of
course, the screen(s) need(s) to support such HD resolution as well.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>