u-boot: colibri-imx7: improve fw_env.config
Rely on BSP fw_env.config for anything but the special eMMC one for this
combined/unified Colibri iMX7 NAND/eMMC machine.

Please note that we first re-name the regular BSP one to
fw_env_mtd4.config as for the BSP this is the NAND machine while the
special eMMC one gets installed as fw_env.config as that is what the
qt-tezi application expects for regular block device based modules.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>