verdin-imx8mp: boot.cmd: force resolution
Force the resolution of all HDMI outputs to 1280x720 by forcing the use
of the now available EDID blob.
This works when one or two HDMI Monitors are connected at boot or when
one or both are hot plugged later.

Without this, if a monitor is connected to the DSI to HDMI connector and
the native HDMI connector is open the monitor is driven at a resolution
of 1080p, but the Toradex Easy Installer application uses only 800x600
part of the screen.

Related-to: TEI-888
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>