wrapup.sh: drop specific apalis-imx6/colibri-imx6 script
For some of the very first Apalis-imx6/Colibri-iMX6 the script
checked if one time fuses in the SoC and eMMC configuration had
to be changed and if so that was done.

These modules no longer exist and the upstream kernel we now use
doesn't provide a SoC fuse functionality.
Drop the specific wrapup.sh script and use the generic, empty one.

Related-to: TEI-896
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>