image_type_tezi_run: create legacy image json files
Toradex Easy Installer versions 2.0b3 and older crash if they install
a Toradex Easy Installer with a version with 3 elements in the image
json file. (e.g. `"version": "5.4.0-devel-20210902193741+build.0"`).

Create a additional image json file which works around that limitation,
so that one can update from 1.8 to e.g. 5.3.0 which might be needed
to then install an image which requires features only available in a
later Toradex Easy Installer version.
Limit that to those machines which do or did ship with such an older
Toradex Easy Installer.

Related-to: TEI-917
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>