export: set use-head-next on integration/next manifest by default
This addition checks if one is using the integration or next manifest
with repo if that is true it will automatically append the correct
use-head-next setting when creating build/conf/local.conf from the default.

Also drop pointing out BB_NUMBER_THREADS, PARALLEL_MAKE as they nowadays
default to number of CPU cores.

Related-to: ELB-4127
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>