recipes-kernel: 5.4-2.3.x: lower mhdp hdmi bridge ddc/edid read delay
The MHDP HDMI bridge driver is implemented in such a way that mostly it does
not check the status of PHY read operations and continuing normal execution even
if the status of the previous read operation is an error. Such an implementation
reduces the boot process speed in case if the hardware setup has a special
configuration e.g. I2C bus used for the display mode configuration is disabled.
Due to a testing need to have a stable boot time, a low-level read timeout
(MAILBOX_TIMEOUT_US) is lowered from 5 seconds to 50 milliseconds.

Signed-off-by: Oleksii Bidnichenko <oleksii.bidnichenko@toradex.com>