qt-tezi: rc.local: start vnc-backend if drm-backend fails
If weston fails starting for any reason we may want to default to
a vnc-only backend.

This has the advantage that qt-tezi application will start as there will
be a compositor running in any way.

There should be no downside doing this as we did rely on the success of
the drm-backend weston-command in the past.

Related-to: TEI-998
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>