u-boot-distro-boot: add verdin-am62 script
Add a script which boots the Toradex Easy Installer either loaded
from eMMC or already loaded to RAM from USB DFU.

Don't implement loading binaries from anywhere but MMC or preloaded
from DFU for now.

For the i.MX based module we changed the boot script name to
boot-tezi.scr as we did not want a U-Boot specific to the Toradex
Easy Installer to boot a BSP image. Mostly to not mix and match
System Controller FW versions with a potentially incompatible API.
For Verdin AM62 the used U-Boot is the same for TEZI and BSP so drop
that distinction.

Related-to: TEI-1066
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>