tezi-run-metadata: am62: recovery: fix error with older dfu-util
Use dfu-util provided with Tezi if the one available in the system does
not support the `-w` option.

This is required since lot of commonly used distribution still ship an
older dfu-util version without support for `-w`, with this change the
script work out of the box in such case.

Related-to: TEI-1081
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>