tezi-run-metadata: am62: recovery: handle soc type
Add support for selecting the correct tiboot3.bin file depending on the
SoC type (GP or HS-FS).

SoC ID is read using dfu from the bootrom, depending on the devType the
correct firmware is selected. The actual format of the SoC ID is
described in this script [1] and it was confirmed by TI FAE.

[1] https://software-dl.ti.com/processor-sdk-linux/esd/AM62X/08_06_00_42/exports/docs/_downloads/parse_uart_boot_socid.py

Related-to: TEI-1083
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>