distro: tezi: Fix UBOOT_ENTRYPOINT address
The UBOOT_ENTRYPOINT variable is set to 0x14200000
by default in meta-toradex-nxp. This value determines
the kernel's loading address by U-Boot, defining the
entry point within tezi.itb.

Change it to 0x11000000, as it is utilized in Tezi 5.
The default value, 0x14200000, is will not work for Tezi
because it conflicts with the default tree file loading location.

Related-to: TEI-1117
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>