tezi-run-metadata: u-boot-distro-boot: Correct file loading addresses
This commit corrects the loading addresses for u-boot.img-recoverytezi,
overlays.txt, and boot.scr files into RAM. The adjustments increase the
spacing between the loaded files and the region where tezi.itb extracts
its images, thereby preventing potential memory corruption errors caused
by file overlap.

The new loading addresses and sizes are shown in the table below:

| File        | RAM Address | Size  | Max Size Available   |
|-------------|-------------|-------|----------------------|
|SPL          |0x910000     |44 KB  |128KB                 |
|u-boot       |0x10000000   |512 KB |All RAM at this point |
|tezi.itb     |0x14420000   |512 KB |64MB                  |
|overlays.txt |0x14410000   |512 KB |64MB                  |
|boot.scr     |0x14400000   |512 KB |64MB                  |

The 'Max Size Available' is calculated between one loading address and
the next one. For example, u-boot will be loaded at the beginning of the
RAM, so it has all the RAM available to him at this point. However,
overlays.txt has only 64MB because tezi.itb will be loaded 0x100000
after it.

'tezi.itb' will also extract the kernel, ramdisk, device tree and device
tree overlays (if it exists) that are packed inside the fit image.
These are the addresses that it will be extracted:

| File   | RAM Address | Size  | Max Size Available |
|--------|-------------|-------|--------------------|
|kernel  |0x11000000   |9 MB   |52 MB               |
|ramdisk |0x1A400000   |26 MB  |95 MB               |
|dtb     |0x14200000   |64 KB  |960 KB              |
|dtbos   |0x142F0000   |-*     |1 MB                |

*For dtbos, the size depends on how much of them we will load.

Related-to: TEI-1117
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>