u-boot-toradex-ti: Use 'toradex-recovery-u-boot.dtsi'
Integrate 'toradex-recovery-u-boot.dtsi' using CONFIG_DEVICE_TREE_INCLUDES
to set 'recovery_tezi' to 1, aligning it with the iMX modules.

By moving the 'do_configure:prepend' to 'u-boot-toradex-tezi.inc', now the
am62 will be configure with the same configs as the iMX modules are
configured (including the CONFIG_DEVICE_TREE_INCLUDES), with the
exception of the BOOTCOMMAND config.

The UBOOT_CONFIG_BASENAME is needed for the verdin-am62-k3r5 as well,
otherwise when Yocto tries to build for the MACHINE it will break the
build.

Additionaly, make minor changes by using spaces instead of tabs and
alphabetically order the configs to ensure consistency with other u-boot
recipes.

Related-to: TEI-1110
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>