tezi: kernel: workaround for GPIOLIB dependency
OE outputs a kernel configuration file without GPIO support,
as the configuration is probably implicitly selected from another one.
This causes runtime dependency problems as some devices are not initialized
because are waiting for the gpio controller. This is a workaround and a
proper solution must be done on BSP.

Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>