linux: tezi.cfg: Disable CDC_COMPOSITE and G_NCM usb configs
`CONFIG_USB_G_NCM` uses the configs to create a `g_ncm` UDC function to
use the ethernet over USB. However, on QT Tezi app, the code will create
a new NCM config called `gncm` by using the libusbgx API, so disable it
here to avoid any issues.

Also, `CONFIG_CDC_COMPOSITE` uses the configs to create e new composite
device called `g_cdc` that has ethernet over usb and serial ACM
communication at the same time. This is not used by Tezi and keeping it
enable prevent the mass storage function `gms` to be correctly loaded,
so this config must be disabled or it needs to be enabled as a module.
So disable it.

Related-to: TEI-1078
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>