tezi: recovery: verdin-am62: Fix windows recovery script
The recovery-windows.bat script has a mistake where it calls uuu and
pauses before completing the remaining dfu steps. This problem arises
because uuu is invoked from within the 'select_tiboot3_bin' subroutine,
causing it to never return and leading to a pause in the downloading
execution. The fix moves uuu call to the correct location.

Related-to: TEI-1150
Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>