tezi: qt5: backport oppenssl3 support
In recent versions of toradex bsp6, the version of openssl in kirkstone
is v3 and it is used to access https:// domains. However qt5 version of
kirkstone meta-qt5 still doesn't have the backport to support
openssl3 in qt5 and this make tezi ui to not work with https:// feeds.
This patch backport the support of openssl3 already on upstream
repository of qtbase to our current version of qt5 until this is
corrected on meta-qt5.

Related-to: TEI-1149
Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>