tezi-run-metadata: Add recovery-windows.README for all machines
Add a README file called recovery-windows.README for all machines
(before it was only added to verdin-am62) with instructions on
how to install the correct ADB driver for a windows host machine. This
will help users with possible errors on uuu fastboot when using it with a
Windows host machine.

Also add instructions about Fastboot protocol for Verdin-am62 and the
windows host machine.

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>