recipes: 2nd pass: cope with moved directory for unpack
Fix the recipes not already done in commit 7b3f8c9edd83 ("recipes:
cope with moved directory for unpack").

OE master (styhead) no longer unpacks in ${WORKDIR} but rather does
it in path defined by the new variable UNPACKDIR.

Cope with the change so that the layer builds against scarthgap and
styhead.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>