recipes: remove OURFILEPATH/S workarounds
The workaround added in commit 7b3f8c9edd83 ("recipes: cope with moved
directory for unpack") is no longer needed. This was initially required to
build this layer with both master (styhead) and scarthgap branches from
other layers, as the meta-toradex-tezi/scarthgap-7.x.y did not exist
at that time. With the creation of a specific scarthgap branch, we can
remove the workarounds.

Related-to: ELB-5840
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>