rc.local: remove screen-share module from backup call
The backup weston command for VNC if the first one fails will not read
the weston.ini file and it will not use screen-share, as this is VNC
only (screen-share is used to mirror the physical ouput with the VNC
backend), therefore this module is not needed. Remove it.

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>
(cherry picked from commit cba1292a96b9d4eaad4970637aef72a5df76a88f)