tezi-run-metadata: k3: uuu.auto: update fastboot buffer address
Since commit 69af2e83fa1e ("verdin-am62: fix FASTBOOT_BUF_ADDR and
FASTBOOT_BUF_SIZE") in U-Boot toradex_ti-u-boot-2024.04, the fastboot buffer
address was updated to work with 512MB and 1GB devices, as the previous
value was not working.

Propagate this change to uuu.auto to fix TEZI not being able to load on
512MB and 1GB devices.

Note: overlays.txt and fastboot buffer address are the same, in other
words, the file does not need to be moved to another RAM address as
tezi.itb and boot.scr. The address is reused.

Related-to: TEI-1223
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>