tezi.conf: remove linux-firmware packages
These packages are being added to Aquila AM69 and Verdin AM62, but they
are not needed for Tezi, as it does not support for Wi-Fi, GPU
accelaration or Bluetooth. These are currently increasing the size of
the image significantly. Remove them to save some space.

Note: linux-firmware-ath12k is not being added as of today, but it will
be in the near future. So let's remove it to prevent this package from
being added.

Related-to: TEI-1232
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>
(cherry picked from commit 0a0d14277202fd8646f59e86fc05e669e40650a0)