verdin-imx8mm: boot.cmd.in: fix video output when connector is disconnected
Commit 7d420ac174e4 introduced a regression in the video kernel
argument, causing it to fail when the screen is not physically
connected. This issue makes Weston fail to initialize:

[13:19:52.140] DRM: head 'HDMI-A-1' found, connector 37 is disconnected.
[13:19:52.146] Cannot pick output: Pointer not on any output, or no
focused/default output found

To fix this, the video argument must include the D option to force
the output to always be connected. Additionally, the frame rate (60) and
output name (HDMI-A-1) must be passed.

This makes sure Weston can initialize the output even when no physical
connection is present, making VNC output work without any display.

Related-to: TEI-1243
Fixes: 7d420ac174e4 ("u-boot-distro-boot: verdin-imx8mm: boot.cmd.in: update kernel arguments")
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>