qt-tezi: rc.local: wait for graphics card before starting weston
Weston may fail to start if the graphics card is not yet ready, this
issue was observed on Verdin iMX8MP.
Add a loop to wait for /dev/dri/card0 and, if /dev/dri/card0 is a GPU,
also wait for /dev/dri/card1 before starting Weston.
This follows the same approach used in the weston.service systemd unit
from the reference image.

Related-to: TEI-1282
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>