u-boot-distro-boot: drop KERNEL_DTB_PREFIX
Master no longer has a prefix for the configuration nodes in a
fitimage. Remove the @@KERNEL_DTB_PREFIX@@ placeholder in the
boot.cmd.in file.

Otherwise the placeholder is not replaced and the resulting boot.cmd
contains an not existing configuration to load and the resulting TEZI
does not boot.

Fixes: commit e2fed1009a3b ("u-boot-distro-boot: unify tezi boot scripts")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>